# static-assets/openapi.yaml
openapi: 3.1.0
info:
  title: Investor Search API
  description: An API that allows users to search for investors.
  version: "v1"
servers:
  - url: http://localhost:5003
paths:
...
  /sapi/instrumentsmetadata/V1.1/instruments/{instrumentId}:
    get:
      operationId: getInstrument
      summary: Get Instrument base info
      parameters:
        - in: path
          name: instrumentId
          schema:
            type: number
          required: true
          minimum: 1
          maximum: 99999999
          description: The id of the instrument if not exists will return all instruments metadata
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/getInstrumentResponse"
...
  /sapi/candles/candles/desc.json/{period}/{candlesNumber}/{instrumentId}:
    get:
      operationId: getInstrumentsCandles
      summary: get instruments prices using candles data including range low,high,open and close
      parameters:
        - in: path
          name: period
          schema:
            type: string
            enum:
              [
                "OneMinute",
                "FiveMinutes",
                "TenMinutes",
                "FifteenMinutes",
                "OneHour",
                "OneDay",
                "OneWeek",
              ]
          required: true
          description: the time frame of the candles
        - name: candlesNumber
          in: path
          description: amount of candles to return
          required: true
          schema:
            type: number
            default: 30
        - name: instrumentId
          in: path
          description: etoro instrument id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/candlesResponse"
  #/sapi/instrumentsinfo/instruments/{instrumentId}:
...
components:
  schemas:
    getInstrumentResponse:
      type: object
      properties:
        InstrumentDisplayData:
          type: object
          properties:
            InstrumentID:
              type: integer
            InstrumentDisplayName:
              type: string
            InstrumentTypeID:
              type: integer
            ExchangeID:
              type: integer
            Images:
              type: array
              items:
                type: object
                properties:
                  InstrumentID:
                    type: integer
                  Width:
                    type: number
                  Height:
                    type: number
                  Uri:
                    type: string
                  BackgroundColor:
                    type: string
                  TextColor:
                    type: string
            SymbolFull:
              type: string
            StocksIndustryID:
              type: integer
            PriceSource:
              type: string
            HasExpirationDate:
              type: boolean
            IsInternalInstrument:
              type: boolean
          description: The instrument display data.
...
    candlesResponse:
      type: object
      properties:
        Candles:
          type: array
          items:
            $ref: "#/components/schemas/Candle"
        Interval:
          type: string
    Candle:
      type: object
      properties:
        Close:
          type: number
        FromDate:
          format: date-time
        High:
          type: number
        InstrumentID:
          type: number
        Low:
          type: number
        Open:
          type: number
...